Adding exception handling, fixed bug in modify reservation

+++ src/main/java/com/volcano/campsite/availability/AvailabilityServiceImpl.java
@@ -1,6 +1,9 @@
 package com.volcano.campsite.availability;
 
 import com.volcano.campsite.common.Helper;
+import com.volcano.campsite.config.CampsiteConfig;
+import com.volcano.campsite.exception.CampsiteErrorCode;
+import com.volcano.campsite.exception.CampsiteException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -19,19 +22,23 @@ public class AvailabilityServiceImpl implements AvailabilityService {
     @Autowired
     private CampgroundAvailabilityRepository campgroundAvailabilityRepository;
 
+    @Autowired
+    private CampsiteConfig campsiteConfig;
+
     @Override
     public List<LocalDate> getAvailableDatesInRange(LocalDate startDate, LocalDate endDate) {
+        logger.info("MAX Campgrounds = {}", campsiteConfig.getMaxCampgrounds());
         if (startDate == null) {
-            startDate = LocalDate.now().plusDays(1);
+            startDate = LocalDate.now().plusDays(campsiteConfig.getDefaultStartDayOffset());
         }
         if (endDate == null) {
-            endDate = startDate.plusDays(30);
+            endDate = startDate.plusDays(campsiteConfig.getDefaultAvailabilityRangeDays());
         }
         List<LocalDate> allDatesInRange = Helper.getAllDatesInRange(startDate, endDate);
         logger.info("Searching for dates = {}", allDatesInRange);
         List<CampgroundAvailabilityEntity> bookedCampgrounds = campgroundAvailabilityRepository.findBookingsForDates(allDatesInRange);
         logger.info("Campgrounds Booked = {}", bookedCampgrounds);
-        bookedCampgrounds.removeIf(campground -> campground.getReservationIds() != null && campground.getReservationIds().size() < 250);
+        bookedCampgrounds.removeIf(campground -> campground.getReservationIds() != null && campground.getReservationIds().size() < campsiteConfig.getMaxCampgrounds());
         bookedCampgrounds.forEach(bookedCampground -> {
             allDatesInRange.removeIf(date -> bookedCampground.getOccupiedDate().equals(date));
         });
@@ -44,8 +51,8 @@ public class AvailabilityServiceImpl implements AvailabilityService {
         List<LocalDate> reservationDates = Helper.getAllDatesInRange(arrivalDate, departureDate);
         List<CampgroundAvailabilityEntity> entityList = campgroundAvailabilityRepository.findBookingsForDates(reservationDates);
         entityList.forEach(entity -> {
-            if(entity.getReservationIds().size() > 250) {
-                logger.info("Throw exception here - dates not available");
+            if(entity.getReservationIds().size() > campsiteConfig.getMaxCampgrounds()) {
+                throw new CampsiteException(CampsiteErrorCode.DATES_NOT_AVAILABLE);
             }
             entity.getReservationIds().add(reservationId);
             reservationDates.removeIf(reservationDate -> entity.getOccupiedDate().equals(reservationDate));
@@ -68,12 +75,16 @@ public class AvailabilityServiceImpl implements AvailabilityService {
 
     @Override
     public void deleteReservationId(UUID reservationId) {
+        logger.info("RESERVATION ID TO BE DELETED = {}", reservationId);
         List<CampgroundAvailabilityEntity> reservedCampgroundsList = campgroundAvailabilityRepository.findBookingsByReservationId(reservationId);
+        if(reservedCampgroundsList.isEmpty()) {
+            throw new CampsiteException(CampsiteErrorCode.RESERVATION_NOT_FOUND);
+        }
         reservedCampgroundsList.forEach(entity -> {
                 entity.getReservationIds().removeIf(id -> id.equals(reservationId));
                 if(entity.getReservationIds().size() == 0) {
                     campgroundAvailabilityRepository.deleteOccupiedDate(entity.getPersistenceId());
-                    reservedCampgroundsList.remove(entity);
+//                    reservedCampgroundsList.remove(entity);
                 }
         });
         campgroundAvailabilityRepository.saveAll(reservedCampgroundsList);
